==> 8_current_maps.Rmd <==
---
title: "Current Map Design"
author: "Dana Roemling"
date: '2023-04-03'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r libs n data}
library(sf)
library(sp)
library(tidyverse)
library(gstat)
library(stringr) # regex
library(dplyr) # regex
library(tidyr)
library(scales)
library(classInt)
library(viridis)
library(viridisLite)
library(rnaturalearth)
library(rnaturalearthhires)

# data on local
corpus <- read.csv(file = './data_ling/full_matrix_for_filtering.csv') 
token_at_location <- read.csv(file = './data_ling/tokens_at_location.csv')
colnames(token_at_location) <- c("City", "Tokencount") 
longlat <- read.csv(file = './data_maps/gsa_geo_filtered.csv')
longlat_small <- read.csv(file = './data_maps/gsa_geo_filtered_nosmall22.csv')
longlat_small <- read.csv(file = './data_maps/gsa_geo_filtered_nosmall85.csv')

# data on server
corpus <- read.csv(file = '/rds/projects/g/grievej-german-dialect-profiling/full_matrix_for_filtering.csv') 
token_at_location <- read.csv(file = '/rds/projects/g/grievej-german-dialect-profiling/tokens_at_location.csv')
colnames(token_at_location) <- c("City", "Tokencount") 
longlat <- read.csv(file = '/rds/projects/g/grievej-german-dialect-profiling/gsa_geo_filtered.csv')
longlat_small <- read.csv(file = '/rds/projects/g/grievej-german-dialect-profiling/gsa_geo_filtered_nosmall22.csv')
longlat_small <- read.csv(file = '/rds/projects/g/grievej-german-dialect-profiling/gsa_geo_filtered_nosmall85.csv')
```

```{r cities}
cities <- data.frame(
  City = c("Cologne", "Munich", "Vienna", "Zurich", "Berlin", "Hamburg"),
  Long = c(6.9578, 11.5755, 16.3731, 8.5417, 13.3833, 10),
  Lat = c(50.9422, 48.1372, 48.2083, 47.3769, 52.5167, 53.55))
crs2 <- CRS("+init=epsg:4326")
cities_sf <- st_as_sf(cities, coords = c("Long", "Lat"), crs = crs2)
```

```{r gsa shape}
# shape file on local
EU <- st_read(dsn="./data_maps/EU/Try/", layer="NUTS_RG_20M_2021_3035")

# shape file on server
EU <- st_read(dsn="/rds/projects/g/grievej-german-dialect-profiling/shape/EU/Try/", layer="NUTS_RG_20M_2021_3035")

# change projection
EU <- st_transform(EU, CRS("+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"))

# get gsa outline
#gsa_string <- c("DE", "AT", "CH")
#gsa <- filter(EU, CNTR_CODE %in% gsa_string)
#gsa_outline <- filter(gsa, LEVL_CODE == 0)
#gsa_plot <- gsa_outline %>% dplyr::select(geometry) # just the outlines of the three
#gsa_spatial <- as_Spatial(gsa_plot) # new object as spatial instead of sf

# large grid
sp_grid <- as.data.frame(spsample(gsa_spatial,
                                  n = 100000,
                                  type = "regular", # systematically aligned sampling
                                  offset = c(0.5,0.5))) # makes sure grid is the same every time
# This grid needs A CRS and to be turned into a spatial object, which this line does:
sp_grid_sf <- st_as_sf(sp_grid, coords=c("x1","x2"), crs = st_crs(4326))



# small grid
sp_grid_small <- as.data.frame(spsample(gsa_spatial,
                                  n = 500,
                                  type = "regular", # systematically aligned sampling
                                  offset = c(0.5,0.5))) # makes sure grid is the same every time
# This grid needs A CRS and to be turned into a spatial object, which this line does:
sp_grid_sf_small <- st_as_sf(sp_grid_small, coords=c("x1","x2"), crs = st_crs(4326))

# with usual code / alternative
gsa_outline <- ne_countries(country = c("Austria", "Germany", "Switzerland"), returnclass="sf", scale = "large")
gsa_plot <- gsa_outline %>% dplyr::select(geometry) # just the outlines of the three
gsa_spatial <- as_Spatial(gsa_plot) # new object as spatial instead of sf
```

```{r counting / regex}
#get word of interest
one_word <- corpus %>% filter(word == "mistkübel")
colnames(one_word) <- c("Token", "City", "Frequency")

#merge with geo data
merger_one <- merge(one_word, longlat, by.x ="City", by.y = "City")
colnames(merger_one) <- c("City", "Token", "Frequency", "lon", "lat")
merger <- merge(merger_one, token_at_location, by.x ="City", by.y = "City")

# OR
# zero inclusion
new_join <- merge(longlat_small, one_word, by.x ="City", by.y = "City", all = TRUE)
new_join <- new_join %>% dplyr::select("City", "lon", "lat", "Frequency")
new_join <- new_join %>% drop_na(lon)
new_join[is.na(new_join)] <- 0
merger <- merge(new_join, token_at_location, by.x ="City", by.y = "City", all.x = TRUE)


# add relative frequency
merger$relfreq <- (merger$Frequency/merger$Tokencount)
merger$relfreq1000 <- (merger$relfreq*1000)
colnames(merger) <- c("City", "Token", "Frequency", "lon", "lat", "TokenCount", "RelativeFrequency", "RelativeFrequencyThousand")
#colnames(merger) <- c("City", "lon", "lat", "Frequency", "TokenCount", "RelativeFrequency", "RelativeFrequencyThousand")
summary(merger$TokenCount)
merger <- merger %>% filter(TokenCount > 1456)

# alternations
one_word <- corpus %>% filter(word == "abitur")
colnames(one_word) <- c("Token", "City", "Frequency")
second_word <- corpus %>% filter(word == "matura")
colnames(second_word) <- c("Token", "City", "Frequency")

# combine into one and calculate proportion
merged <- merge(one_word, second_word, by ="City", all=T)
# introduce zeros for NAs
merged$Frequency.x[is.na(merged$Frequency.x)] <- 0
merged$Frequency.y[is.na(merged$Frequency.y)] <- 0
merged$Proportion <- (merged$Frequency.x / (merged$Frequency.x + merged$Frequency.y) )
merged$Percentage <- 100* (merged$Frequency.x / (merged$Frequency.x + merged$Frequency.y) )

# merge with geo data
merger <- merge(merged, longlat, by.x ="City", by.y = "City")
```

```{r single value variogram}
# create the data frame needed for variogram
trial <- merger %>% dplyr::select("lon", "lat", "RelativeFrequencyThousand") #%>% na.omit()
data <- trial[3]
coords <- trial[1:2]
crs <- CRS("+init=epsg:4326") # Different way of setting same CRS
#crs <- CRS("+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0")
spdf <- SpatialPointsDataFrame(coords      = coords,
                               data        = data, 
                               proj4string = crs)

#need to remove duplicates
spdf = spdf[which(!duplicated(spdf@coords)), ]

#st_crs(spdf)
#coordinates(vario_data) <- ~lon+lat # this does the same to turn it into spdf object

vg <- variogram(RelativeFrequencyThousand~1, data = spdf, width = 1, cutoff = 300)
# width seems to limit my analysis, without it makes the vg easier. 

vg_fit <- fit.variogram(vg, vgm("Exp")) 
plot(vg, vg_fit)
```

```{r kriging}
krig1 <- krige(RelativeFrequencyThousand~1, 
               locations = spdf, 
               newdata = sp_grid_sf, 
               model = vg_fit)
```

```{r smoothing and kriging}
krig1 <- krige(RelativeFrequencyThousand~1, 
               locations = spdf, 
               newdata = sp_grid_sf_small, 
               model = vg_fit)

krig2 <- krige(var1.pred~1, 
               locations = krig1, 
               newdata = sp_grid_sf, 
               model = vg_fit)
```

```{r double value / comparison variogram & kriging}
# for vg and krig
trial <- merger %>% dplyr::select("lon", "lat", "Proportion") #%>% na.omit()
data <- trial[3]
coords <- trial[1:2]
crs <- CRS("+init=epsg:4326") # Different way of setting same CRS
spdf <- SpatialPointsDataFrame(coords = coords, data = data, proj4string = crs)
spdf = spdf[which(!duplicated(spdf@coords)), ]
vg <- variogram(Proportion~1, data = spdf, width = 1, cutoff = 300)
vg_fit <- fit.variogram(vg, vgm("Exp")) 
krig_comp_1 <- krige(Proportion~1, locations = spdf, newdata = sp_grid_sf_small, model = vg_fit)
krig_comp_2 <- krige(var1.pred~1, locations = krig_comp_1, newdata = sp_grid_sf, model = vg_fit)
```


```{r legend text single plot}
legend_max <- round(max(krig2$var1.pred), digits = 2)
legend_min <- round(min(krig2$var1.pred), digits = 2)
```

```{r optional text annotation}
# Top:
annotate(geom="text", x = 11, y = 55.5, label="Diese Karte zeigt das Verteilung des Wortes “bummeln” \nin Deutschland, Österreich und der Schweiz. \nStärker gefärbte Regionen nutzen das Wort mehr.", size = 2, family = "Optima") 
annotate(geom="text", x = 11, y = 55.5, label="Diese Karte zeigt das Verhältnis zwischen “schau” und “guck” \nin Deutschland, Österreich und der Schweiz. \nRegionen in lila nutzen mehr “schau” und Regionen in orange nutzen mehr “guck”.", size = 2, family = "Optima")
# Bottom:
# NB: Needs adjusting of legend points!! Use 51.1 and 52.2 for heights - -0.2
annotate(geom="text", x = 11, y = 45.2, label="Daten aus der App Jodel im Zeitraum April - Juni 2017\n21 Mio. Posts gesammelt von Hovy & Purschke 2018\nKriging der relativen Häufigkeiten", size = 2, family = "Optima") 
```


```{r light mode plot cont fill purple}
ggplot() +
  geom_sf(data = krig2, aes(fill=var1.pred), shape = 21, size = 0.5, stroke = 0, lwd = 0) +
  geom_sf(data = gsa_plot, aes(geometry = geometry), color="black", fill=NA, size = 0.5) +
  geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, family = "Optima") +
  geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4) +
  theme_minimal() +
  scale_fill_continuous(low="#ffffff", high="#310d59", name = "'nicht'", labels = NULL) +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        panel.grid.major = element_blank()) +
  theme(legend.position = c(0.90, 0.65), 
        legend.title = element_text(size= 6, family = "Optima"), 
        legend.text = element_text(size = 6, family = "Optima", hjust = 1),
        legend.title.align = 0,
        legend.key.size = unit(0.3, "cm"),
        legend.key.width = unit(0.4,"cm")) +
  annotate(geom="text", x = 16.6, y = 52.4, label=legend_max, size = 1.5, family = "Optima") +
  annotate(geom="text", x = 16.6, y = 51.3, label=legend_min, size = 1.5, family = "Optima") +
  annotate(geom="text", x = 16.2, y = 49.8, label="danaroemling.com", size = 2,5, family = "Optima") +
  annotate(geom="text", x = 16.2, y = 49.6, label="03/04/2023", size = 2,5, family = "Optima")
```

```{r light mode plot 7 bands purple}
pal_light_1 <- c("#ffffff", "#dcd3e2", "#b9a8c5", "#977fa9",  "#75588e", "#533373", "#310d59")
ggplot() +
  geom_sf(data = krig1, aes(fill=var1.pred), shape = 21, size = 0.5, stroke = 0, lwd = 0) +
  geom_sf(data = gsa_plot, aes(geometry = geometry), color="black", fill=NA, size = 0.5) +
  geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, family = "Optima") +
  geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4) +
  theme_minimal() +
  scale_fill_stepsn(colours = pal_light_1, 
                    breaks = seq(min(krig3$var1.pred), max(krig3$var1.pred), length = 8),
                    values = seq(0.5/14, 13/14, length = 7),
                    labels = NULL) +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        panel.grid.major = element_blank()) +
  theme(legend.position = c(0.90, 0.65), 
        legend.title = element_text(size = 6, family = "Optima"), 
        legend.text = element_text(size = 6, family = "Optima", hjust = 1),
        legend.title.align = 0,
        legend.key.size = unit(0.3, "cm"),
        legend.key.width = unit(0.4,"cm")) +
  annotate(geom="text", x = 16.6, y = 52.4, label=legend_max, size = 1.5, family = "Optima") +
  annotate(geom="text", x = 16.6, y = 51.3, label=legend_min, size = 1.5, family = "Optima") +
  annotate(geom="text", x = 16.2, y = 49.8, label="danaroemling.com", size = 2.5, family = "Optima") +
  annotate(geom="text", x = 16.2, y = 49.6, label="03/04/2023", size = 2.5, family = "Optima") +
  labs(fill = "Mehr 'nichts'")
```

```{r dark mode plot cont fill green}
ggplot() +
  geom_sf(data = krig1, aes(fill = var1.pred), shape = 21, size = 0.5, stroke = 0, lwd = 0) +
  geom_sf(data = gsa_plot, aes(geometry = geometry), color = "snow2", fill = NA, size = 0.5) +
  geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, color = "snow", family = "Optima") +
  geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4, color = "snow") +
  theme_minimal() +
  scale_fill_continuous(low="#1b1b1b", high="#73ffc2", name = "Mehr 'nicht'", labels = NULL) +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        panel.grid.major = element_blank()) +
  theme(legend.position = c(0.90, 0.65), 
        legend.title = element_text(size = 7, color = "snow", family = "Optima"), 
        legend.text = element_text(size = 6, color = "snow", family = "Optima", hjust = 1),
        legend.title.align = 0,
        panel.background = element_rect(fill = "gray10", color = "gray10"),
        plot.background = element_rect(fill = "gray10", color = "gray10"),
        legend.key.size = unit(0.3, "cm"),
        legend.key.width = unit(0.4,"cm")) +
  annotate(geom="text", x = 16.6, y = 52.4, 
           label=legend_max, size = 1.5, family = "Optima", color = "snow") +
  annotate(geom="text", x = 16.6, y = 51.3, 
           label=legend_min, size = 1.5, family = "Optima", color = "snow") +
  annotate(geom="text", x = 16.2, y = 49.8, 
           label="danaroemling.com", size = 2.5, color = "snow", family = "Optima") +
  annotate(geom="text", x = 16.2, y = 49.6, 
           label="28/02/2023", size = 2.5, color = "snow", family = "Optima")
```

```{r dark mode plot 7 bands green}
pal_new_green <- c("#1b1b1b", "#2b3b33", "#3a5e4d", "#498468", "#57ab85", "#65d4a3", "#73ffc2")
ggplot() +
  geom_sf(data = krig3, aes(fill = var1.pred), shape = 21, size = 0.5, stroke = 0, lwd = 0) +
  geom_sf(data = gsa_plot, aes(geometry = geometry), color = "snow2", fill = NA, size = 0.5) +
  geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, color = "snow", family = "Optima") +
  geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4, color = "snow") +
  theme_minimal() +
  scale_fill_stepsn(colours = pal_new_green, 
                    breaks = seq(min(krig3$var1.pred), max(krig3$var1.pred), length = 8),
                    values = seq(0.5/14, 13/14, length = 7),
                    labels = NULL) +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        panel.grid.major = element_blank()) +
  theme(legend.position = c(0.90, 0.65), 
        legend.title = element_text(size = 6, color = "snow", family = "Optima"), 
        legend.text = element_text(size = 6, color = "snow", family = "Optima", hjust = 1),
        legend.title.align = 0,
        panel.background = element_rect(fill = "gray10", color = "gray10"),
        plot.background = element_rect(fill = "gray10", color = "gray10"),
        legend.key.size = unit(0.3, "cm"),
        legend.key.width = unit(0.4,"cm")) +
  annotate(geom="text", x = 16.6, y = 52.4, 
           label=legend_max, size = 1.5, family = "Optima", color = "snow") +
  annotate(geom="text", x = 16.6, y = 51.3, 
           label=legend_min, size = 1.5, family = "Optima", color = "snow") +
  annotate(geom="text", x = 16.2, y = 49.8, 
           label="danaroemling.com", size = 2.5, color = "snow", family = "Optima") +
  annotate(geom="text", x = 16.2, y = 49.6, 
           label="20/02/2023", size = 2.5, color = "snow", family = "Optima") +
  labs(fill = "Mehr 'nichts'")
```

```{r dark mode isogloss plot green}
summary(krig3$var1.pred)
ggplot() +
  geom_sf(data = krig3, aes(fill = var1.pred), shape = 21, size = 0.5, stroke = 0, lwd = 0) +
  geom_sf(data = gsa_plot, aes(geometry = geometry), color = "snow2", fill = NA, size = 0.5) +
  geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, color = "snow", family = "Optima") +
  geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4, color = "snow") +
  theme_minimal() +
  # this needs adjusting at breaks below
  scale_fill_steps(low="#1b1b1b", high="#73ffc2", breaks = 10.9, labels = NULL) +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        panel.grid.major = element_blank()) +
  theme(legend.position = c(0.90, 0.65), 
        legend.title = element_text(size = 6, color = "snow", family = "Optima"), 
        legend.text = element_text(size = 6, color = "snow", family = "Optima", hjust = 1),
        legend.title.align = 0,
        panel.background = element_rect(fill = "gray10", color = "gray10"),
        plot.background = element_rect(fill = "gray10", color = "gray10"),
        legend.key.size = unit(0.3, "cm"),
        legend.key.width = unit(0.4,"cm")) +
  annotate(geom="text", x = 16.6, y = 52.4, 
           label=legend_max, size = 1.5, family = "Optima", color = "snow") +
  annotate(geom="text", x = 16.6, y = 51.3, 
           label=legend_min, size = 1.5, family = "Optima", color = "snow") +
  annotate(geom="text", x = 16.2, y = 49.8, 
           label="danaroemling.com", size = 2.5, color = "snow", family = "Optima") +
  annotate(geom="text", x = 16.2, y = 49.6, 
           label="03/04/2023", size = 2.5, color = "snow", family = "Optima") +
  labs(fill = "Mehr 'nichts'")
```

```{r light mode plot N bands purple}
ggplot() +
  geom_sf(data = krig3, aes(fill=var1.pred), shape = 21, size = 0.75, stroke = 0, lwd = 0) +
  geom_sf(data = gsa_plot, aes(geometry = geometry), color="black", fill=NA, size = 0.5) +
  geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, family = "Optima") +
  geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4) +
  theme_minimal() +
  scale_fill_steps(low="#ffffff", high="#310d59", n.breaks = 30, labels = NULL) +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        panel.grid.major = element_blank()) +
  theme(legend.position = c(0.90, 0.65), 
        legend.title = element_text(size = 6, family = "Optima"), 
        legend.text = element_text(size = 6, family = "Optima", hjust = 1),
        legend.title.align = 0,
        legend.key.size = unit(0.3, "cm"),
        legend.key.width = unit(0.4,"cm")) +
  annotate(geom="text", x = 16.6, y = 52.4, label=legend_max, size = 1.5, family = "Optima") +
  annotate(geom="text", x = 16.6, y = 51.3, label=legend_min, size = 1.5, family = "Optima") +
  annotate(geom="text", x = 16.2, y = 49.8, label="danaroemling.com", size = 2.5, family = "Optima") +
  annotate(geom="text", x = 16.2, y = 49.6, label="03/04/2023", size = 2.5, family = "Optima") +
  labs(fill = "Mehr 'nichts'")
```

```{r dark mode plot N bands green}
# Change the number of n.breaks = to adjust number of break points
ggplot() +
  geom_sf(data = krig3, aes(fill = var1.pred), shape = 21, size = 1, stroke = 0, lwd = 0) +
  geom_sf(data = gsa_plot, aes(geometry = geometry), color = "snow2", fill = NA, size = 0.5) +
  geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, color = "snow", family = "Optima") +
  geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4, color = "snow") +
  theme_minimal() +
  scale_fill_steps(low="#1b1b1b", high="#73ffc2", n.breaks = 30, labels = NULL) +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        panel.grid.major = element_blank()) +
  theme(legend.position = c(0.90, 0.65), 
        legend.title = element_text(size = 6, color = "snow", family = "Optima"), 
        legend.text = element_text(size = 6, color = "snow", family = "Optima", hjust = 1),
        legend.title.align = 0,
        panel.background = element_rect(fill = "gray10", color = "gray10"),
        plot.background = element_rect(fill = "gray10", color = "gray10"),
        legend.key.size = unit(0.3, "cm"),
        legend.key.width = unit(0.4,"cm")) +
  annotate(geom="text", x = 16.6, y = 52.4, 
           label=legend_max, size = 1.5, family = "Optima", color = "snow") +
  annotate(geom="text", x = 16.6, y = 51.3, 
           label=legend_min, size = 1.5, family = "Optima", color = "snow") +
  annotate(geom="text", x = 16.2, y = 49.8, 
           label="danaroemling.com", size = 2.5, color = "snow", family = "Optima") +
  annotate(geom="text", x = 16.2, y = 49.6, 
           label="03/04/2023", size = 2.5, color = "snow", family = "Optima") +
  labs(fill = "Mehr 'nichts'")
```


```{r dark mode plot N bands red}
# Change the number of n.breaks = to adjust number of break points
ggplot() +
  geom_sf(data = krig2, aes(fill = var1.pred), shape = 21, size = 0.75, stroke = 0, lwd = 0) +
  geom_sf(data = gsa_plot, aes(geometry = geometry), color = "snow2", fill = NA, size = 0.5) +
  geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, color = "snow", family = "Optima") +
  geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4, color = "snow") +
  theme_minimal() +
  scale_fill_steps(low="#000000", high="firebrick1", n.breaks = 30, labels = NULL) +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        panel.grid.major = element_blank()) +
  theme(legend.position = c(0.90, 0.65), 
        legend.title = element_text(size = 6, color = "snow", family = "Optima"), 
        legend.text = element_text(size = 6, color = "snow", family = "Optima", hjust = 1),
        legend.title.align = 0,
        panel.background = element_rect(fill = "#303030", color = "#303030"),
        plot.background = element_rect(fill = "#303030", color = "#303030"),
        legend.key.size = unit(0.3, "cm"),
        legend.key.width = unit(0.4,"cm")) +
  annotate(geom="text", x = 16.6, y = 52.4, 
           label=legend_max, size = 1.5, family = "Optima", color = "snow") +
  annotate(geom="text", x = 16.6, y = 51.3, 
           label=legend_min, size = 1.5, family = "Optima", color = "snow") +
  labs(fill = "More 'aber'")
```



```{r legend text comparison plot}
legend_max_c <- round(max(krig_comp_2$var1.pred), digits = 2)
legend_min_c <- round(min(krig_comp_2$var1.pred), digits = 2)
```

```{r dark mode plot comparison red blue}
ggplot() +
    geom_sf(data = krig_comp_2, aes(fill=var1.pred), shape = 21, size = 0.5, stroke = 0, lwd = 0) +
    geom_sf(data = gsa_plot, aes(geometry = geometry), color = "snow2", fill = NA, size = 0.5) +
    geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, family = "Optima", color="snow2") +
    geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4, color="snow2") +
    theme_minimal() +
    scale_fill_steps2(low = "blue", mid = "#303030", high = "firebrick1", midpoint = 0.5, n.breaks = 15, labels = NULL) +
    theme(axis.title.x = element_blank(), 
          axis.title.y = element_blank(),
          axis.text.x = element_blank(),
          axis.text.y = element_blank(),
          panel.grid.major = element_blank()) +
    theme(legend.position = c(0.90, 0.65), 
          legend.title = element_text(size = 6, family = "Optima", color = "snow"),
          legend.text = element_text(size = 6, family = "Optima", hjust = 1, color = "snow"),
          legend.title.align = 0,
          legend.key.size = unit(0.3, "cm"),
          legend.key.width = unit(0.4,"cm"),
          panel.background = element_rect(fill = "#303030", color = "#303030"),
          plot.background = element_rect(fill = "#303030", color = "#303030"),
          plot.title = element_text(hjust = 0.5, color = "snow", family = "Optima")) +
    annotate(geom="text", x = 16.6, y = 52.4, label=legend_max_c, size = 1.5, family = "Optima", color = "snow") +
    annotate(geom="text", x = 16.6, y = 51.3, label=legend_min_c, size = 1.5, family = "Optima", color = "snow") +
    labs(fill = "Schau / Guck")
```

```{r light mode comparison orange purple}
summary(krig_comp_2$var1.pred)
ggplot() +
    geom_sf(data = krig_comp_2, aes(fill=var1.pred), shape = 21, size = 0.5, stroke = 0, lwd = 0) +
    geom_sf(data = gsa_plot, aes(geometry = geometry), color="black", fill=NA, size = 0.5) +
    geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, family = "Optima") +
    geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4) +
    theme_minimal() +
    #scale_fill_steps2(low = "#E0961A", mid = "#ffffff", high = "#310d59", midpoint = 0.5, n.breaks = 15, labels = NULL) +
  scale_fill_steps2(low = "seagreen4", mid = "#ffffff", high = "mediumpurple4", midpoint = 0.5, n.breaks = 15, labels = NULL) +
    theme(axis.title.x = element_blank(), 
          axis.title.y = element_blank(),
          axis.text.x = element_blank(),
          axis.text.y = element_blank(),
          panel.grid.major = element_blank()) +
    theme(legend.position = c(0.90, 0.65), 
          legend.title = element_text(size = 6, family = "Optima"),
          legend.text = element_text(size = 6, family = "Optima", hjust = 1),
          legend.title.align = 0,
          legend.key.size = unit(0.3, "cm"),
          legend.key.width = unit(0.4,"cm")) +
    annotate(geom="text", x = 16.6, y = 52.4, label=legend_max_c, size = 1.5, family = "Optima") +
    annotate(geom="text", x = 16.6, y = 51.3, label=legend_min_c, size = 1.5, family = "Optima") +
    annotate(geom="text", x = 16.2, y = 49.8, label="danaroemling.com", size = 2.5, family = "Optima") +
    annotate(geom="text", x = 16.2, y = 49.6, label="03/04/2023", size = 2.5, family = "Optima") +
    labs(fill = "Schau / Guck")
```

```{r light mode comparison orange purple legend w word}
summary(krig_comp_2$var1.pred)
ggplot() +
    geom_sf(data = krig_comp_2, aes(fill=var1.pred), shape = 21, size = 0.5, stroke = 0, lwd = 0) +
    geom_sf(data = gsa_plot, aes(geometry = geometry), color="black", fill=NA, size = 0.5) +
    geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, family = "Optima") +
    geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4) +
    theme_minimal() +
    scale_fill_steps2(low = "#E0961A", mid = "#ffffff", high = "#310d59", midpoint = 0.5, n.breaks = 15, labels = NULL) +
    theme(axis.title.x = element_blank(), 
          axis.title.y = element_blank(),
          axis.text.x = element_blank(),
          axis.text.y = element_blank(),
          panel.grid.major = element_blank()) +
    theme(legend.position = c(0.90, 0.65), 
          legend.title = element_text(size = 6, family = "Optima"),
          legend.text = element_text(size = 6, family = "Optima", hjust = 1),
          legend.title.align = 0,
          legend.key.size = unit(0.3, "cm"),
          legend.key.width = unit(0.4,"cm")) +
    annotate(geom="text", x = 16.6, y = 52.4, label="guck", size = 1.5, family = "Optima") +
    annotate(geom="text", x = 16.6, y = 51.3, label="schau", size = 1.5, family = "Optima") +
    annotate(geom="text", x = 16.2, y = 49.8, label="danaroemling.com", size = 2.5, family = "Optima") +
    annotate(geom="text", x = 16.2, y = 49.6, label="03/04/2023", size = 2.5, family = "Optima") +
    labs(fill = "Distribution")
```

```{r dark mode comparison green yellow}
summary(krig_comp_2$var1.pred)
ggplot() +
    geom_sf(data = krig_comp_2, aes(fill=var1.pred), shape = 21, size = 0.5, stroke = 0, lwd = 0) +
    geom_sf(data = gsa_plot, aes(geometry = geometry), color="snow2", fill=NA, size = 0.5) +
    geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, color = "snow", family = "Optima") +
    geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4, color = "snow") +
    theme_minimal() +
    scale_fill_steps2(low = "#ECE74F", mid = "#111111", high = "#73ffc2", midpoint = 0.5, n.breaks = 15, labels = NULL) +
    theme(axis.title.x = element_blank(), 
          axis.title.y = element_blank(),
          axis.text.x = element_blank(),
          axis.text.y = element_blank(),
          panel.grid.major = element_blank()) +
    theme(legend.position = c(0.90, 0.65), 
          legend.title = element_text(size = 6, color = "snow", family = "Optima"),
          legend.text = element_text(size = 6, color = "snow", family = "Optima", hjust = 1),
          legend.title.align = 0,
          panel.background = element_rect(fill = "gray10", color = "gray10"),
          plot.background = element_rect(fill = "gray10", color = "gray10"),
          legend.key.size = unit(0.3, "cm"),
          legend.key.width = unit(0.4,"cm")) +
    annotate(geom="text", x = 16.6, y = 52.4, label=legend_max_c, size = 1.5, family = "Optima", color = "snow") +
    annotate(geom="text", x = 16.6, y = 51.3, label=legend_min_c, size = 1.5, family = "Optima", color = "snow") +
    annotate(geom="text", x = 16.2, y = 49.8, label="danaroemling.com", size = 2.5, color = "snow", family = "Optima") +
    annotate(geom="text", x = 16.2, y = 49.6, label="03/04/2023", size = 2.5, color = "snow", family = "Optima") +
    labs(fill = "Schau / Guck")
```

```{r dark mode comparison green yellow legend w word}
summary(krig_comp_2$var1.pred)
ggplot() +
    geom_sf(data = krig_comp_2, aes(fill=var1.pred), shape = 21, size = 0.5, stroke = 0, lwd = 0) +
    geom_sf(data = gsa_plot, aes(geometry = geometry), color="snow2", fill=NA, size = 0.5) +
    geom_sf_text(data = cities_sf, aes(label = City), size = 2.5, nudge_x = 0, nudge_y = -0.15, color = "snow", family = "Optima") +
    geom_sf(data = cities_sf, aes(geometry = geometry), shape = 4, color = "snow") +
    theme_minimal() +
    scale_fill_steps2(low = "#ECE74F", mid = "#111111", high = "#73ffc2", midpoint = 0.5, n.breaks = 15, labels = NULL) +
    theme(axis.title.x = element_blank(), 
          axis.title.y = element_blank(),
          axis.text.x = element_blank(),
          axis.text.y = element_blank(),
          panel.grid.major = element_blank()) +
    theme(legend.position = c(0.90, 0.65), 
          legend.title = element_text(size = 6, color = "snow", family = "Optima"),
          legend.text = element_text(size = 6, color = "snow", family = "Optima", hjust = 1),
          legend.title.align = 0,
          panel.background = element_rect(fill = "gray10", color = "gray10"),
          plot.background = element_rect(fill = "gray10", color = "gray10"),
          legend.key.size = unit(0.3, "cm"),
          legend.key.width = unit(0.4,"cm")) +
    annotate(geom="text", x = 16.6, y = 52.4, label="schau", size = 1.5, family = "Optima", color = "snow") +
    annotate(geom="text", x = 16.6, y = 51.3, label="guck", size = 1.5, family = "Optima", color = "snow") +
    annotate(geom="text", x = 16.2, y = 49.8, label="danaroemling.com", size = 2.5, color = "snow", family = "Optima") +
    annotate(geom="text", x = 16.2, y = 49.6, label="03/04/2023", size = 2.5, color = "snow", family = "Optima") +
    labs(fill = "Distribution")
```
